Add Chef Maker support
Chef uses foodcritic to check for Chef specific linting.

Chef also is a superset of Ruby, but running foodcritic against plain
Ruby does not make sense. In order to get this working I had to make a
"dummy" maker for Ruby.

I read [this issue](https://github.com/neomake/neomake/issues/116) and
noticed that the [pull
request](https://github.com/neomake/neomake/pull/133) that fixes it
works if the makers also support the subset, but not just the superset.